Fixes issues with Sendgrid multiple recipients and global variables.
Sendgrid API does not support mixing Smtpapi methods with non-Smtpapi
methods, therefore sending email to multiple recipients with additional
cc and/or bcc emails was not working. Now, all recipients are treated as
of 'to' type.
Sendgrid sections (in our understanding - global variables) can be used
only with substitutions. In order to allow defining global variables,
'fake' substitutions are created for every recipient that point to
specified section value.